// WeatherApp.Portal/src/app/country.d.ts
export interface Country {
    countryCode: string;
    countryName: string;
    isoNumeric: string;
    isoAlpha3: string;
    fipsCode: string;
    continent: string;
    continentName: string;
    capital: string;
    areaInSqKm: number;
    population: number;
    currencyCode: string;
    languages: string;
    geonameId: number;
    west: number;
    north: number;
    east: number;
    south: number;
    postalCodeFormal: string;
}